requirements.txt
openai>=1.6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
qdrant-client>=1.12.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0